Update limma contrasts
Added option to create diagnostic plots and associated arguments. The updated version of plot_pca_v2 is used for one of the plots. Added more details and a link to the LIMMA User's Guide. Additional arguments are now passed to eBayes instead of lmFit. Added option for weighted least squares with var.group argument and arrayWeights.