Minor import updates
Updated what is being imported from different packages. There is still a note about using triple colons to access a function not exported by stats, but it must be happening in one of the functions loaded with getFromNamespace (either the two here or the one elsewhere). No way to fix that on our end.