Update R-CMD-check.yaml
Use devtools::check instead of rcmdcheck::rcmdcheck. Install heatmap.plus package from cran GitHub repo, since it is no longer in CRAN. Do not run on Ubuntu (I am not familiar with this OS, so I could not fix the errors when trying to install devtools). In check(), cran is set to FALSE so more strict CRAN checks will not be run. Also do not build vignettes because "ontology_overrepresentation.Rmd" fails due to Bioconductor SSL certification error that I am not able to fix. It might resolve on its own, but do not build vignettes for now.